Routing afgewerkt - Player scherm

+++ Crontabs/cleanAPI.php
@@ -24,12 +24,15 @@ while($world = sqlsrv_fetch_array($getWorlds, SQLSRV_FETCH_ASSOC))
     {
         $dates = sqlsrv_query($conn, "SELECT DISTINCT(Date) FROM Player WHERE Server_Name = '$world[Server_Name]' AND World_Id = '$world[Id]' Order By Date Desc");
         $date = sqlsrv_fetch_array($dates, SQLSRV_FETCH_NUMERIC);
-        $lastDate = date("Y-m-d", strtotime($date[0]->format("Y-m-d")));
+        if($date[0])
+        {
+            $lastDate = date("Y-m-d", strtotime($date[0]->format("Y-m-d")));
 
-        sqlsrv_query($conn, "DELETE FROM Town WHERE Date < '$lastDate' AND Server_Name = '$world[Server_Name]' AND World_Id = '$world[Id]'");
-        sqlsrv_query($conn, "DELETE FROM Alliance WHERE Date < '$lastDate' AND Server_Name = '$world[Server_Name]' AND World_Id = '$world[Id]' ");
-        sqlsrv_query($conn, "DELETE FROM Player WHERE Date < '$lastDate' AND Server_Name = '$world[Server_Name]' AND World_Id = '$world[Id]' ");
+            sqlsrv_query($conn, "DELETE FROM Town WHERE Date < '$lastDate' AND Server_Name = '$world[Server_Name]' AND World_Id = '$world[Id]'");
+            sqlsrv_query($conn, "DELETE FROM Alliance WHERE Date < '$lastDate' AND Server_Name = '$world[Server_Name]' AND World_Id = '$world[Id]' ");
+            sqlsrv_query($conn, "DELETE FROM Player WHERE Date < '$lastDate' AND Server_Name = '$world[Server_Name]' AND World_Id = '$world[Id]' ");
  
+        }
     }
 }
